moved type tests inside specific dir
